requirements.txt
python-dotenv==1.0.0
pydantic==2.5.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-core==0.1.0
langchain==0.1.0
langchain-community==0.0.10
ddgs==4.1.0
wikipedia==1.4.0 